src/requirements.txt
selenium==4.25.0
chromedriver-binary==132.0.6798.0.0
python-dotenv==0.14.0
beautifulsoup4==4.10.0
webdriver_manager==4.0.2
jpholiday==0.1.10
line-bot-sdk==3.13.0